Revert "Relocate profitbricks util module"
This reverts commit b0eb041ab1cdf57f0c64f2253a780ce029b1a3b7.
