+++ lib/ansible/modules/cloud/profitbricks/profitbricks_datacenter.py
@@ -108,7 +108,7 @@ except ImportError:
 
 from ansible.module_utils.basic import AnsibleModule, env_fallback
 from ansible.module_utils._text import to_native
-from ansible.module_utils.cloud.profitbricks.profitbricks import (
+from ansible.module_utils.profitbricks import (
     LOCATIONS,
     uuid_match,
     wait_for_completion
